fix: update dependencies to build with go 1.23
Since go 1.10 or so, they've wanted to have a go.mod file, and
juju/replicaset depends on juju/mgo/v2 so update those dependencies.
Similarly hashicorp/raft changed their logging interface, so update to
use the hclog dependency.
It would be good to have an adapter from hclog to loggo, but for now, it
at least compiles and you can try to run the code.