Create snapcraft.yaml
Because the app binary will link dynamically to glibc, care has to be
taken to match glibc version between the build host and the OS on the
controller unit. In particular, there is a breaking change in glibc
around Focal that prevents a binary compiled on a more recent Ubuntu
version from running on an older Ubuntu version, e.g. building on
Oracular, running on Bionic.

The snap provides a consistent run-time environment that prevents this
issue.

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>